pyproject.toml
[project]
name = "exemplo-langfuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.15",
    "langfuse>=2.60.3",
    "langgraph>=0.3.31",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
]
